refactor: moved router path into constant folder
